Separate CLI orchestrator/metadata/artifact into their own files
