+++ src/zenml/cli/__init__.py
@@ -13,9 +13,12 @@
 #  permissions and limitations under the License.
 
 
+from zenml.cli.artifact import *  # noqa
 from zenml.cli.base import *  # noqa
 from zenml.cli.config import *  # noqa
 from zenml.cli.example import *  # noqa
+from zenml.cli.metadata import *  # noqa
+from zenml.cli.orchestrator import *  # noqa
 from zenml.cli.pipeline import *  # noqa
 from zenml.cli.stack import *  # noqa
 from zenml.cli.version import *  # noqa
